LayoutTests/sputnik/Conformance/12_Statement/12.1_Block/S12.1_A4_T2-expected.txt

CONSOLE MESSAGE: line 80: SyntaxError: Expected an identifier but found '}' instead
S12.1_A4_T2

PASS Expected parsing failure

TEST COMPLETE
